app/requirements-dev.txt
django-debug-toolbar==5.0